Levels + Elements Changes
Changed how I created upper-levels to use the same operations as
current-level.
When possible, elements have an additional parameter to determine if
their are to be top-linked or not. Even when modifying the height of an
element they can still be linked to a level.
Curtain walls were enhanced to allow the modification of additional
parameters to control their structure.